bubblesheet_backend/requirements.txt
Django==5.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mongoengine==0.27.0
pymongo==4.6.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
reportlab==4.1.0
Pillow==12.2.1
PyMuPDF==1.23.26
opencv-python==4.12.0
numpy==2.2.6
requests==2.32.3